+++ kitchenassistant-android-app/src/main/java/pl/szczeliniak/kitchenassistant/android/network/responses/Page.kt
@@ -3,6 +3,6 @@ package pl.szczeliniak.kitchenassistant.android.network.responses
 data class Page<T>(
     val pageNumber: Long,
     val pageSize: Int,
-    val totalNumberOfPages: Int,
+    val totalNumberOfPages: Long,
     val items: Collection<T>
 )
